cleaning the build system
created customized environments and applied them to a particular node
instead of trying to create one environment (the outcome
is the same, but should be easier to maintain and
add new functionality), still some redundancy that should
be removed, but later (e.g. the function that tries
to automatically recognize the node is copied in
all SConstruct (likely extra module is required). Compiles
on keeneland.