Added backend interface for process registration.
inotify is used to monitor a 'registration' directory. CUDA applications with
the interposing library will cause a file to be written to this directory
indicating the proccess' intention to register with the assembly runtime. This
file is mmap'd by both the backend and any process performing this registration.
In this shared region exist queue(s) for shuffling marshaled functions and data.