Enabled interposer library to mmap with backend.
The interposer library can only map in one file for now. Later, when more
regions are needed will support for more files be added. Lots of TODOs in the
code illustrate where changes would need to be made. Implementation for the
library functions is unclean at the moment, but will be cleaned up later.