Initial commit for shmgrp.
shmgrp is intended to replace libregistration. It is more flexible, allowing one
process to simultaneously be the leader of zero or more process groups and join
zero or more process groups. A member may share zero or more memory regions with
its group leader, and a group leader may share zero or more memory regions with
each member in a group it leads.