Temporary preprocessor macros to control forking.
Definining HACK_NOFORK prevents node initialization in the assembly runtime from
forking to call into the CUDA Runtime API to determine GPU information.

Defining LOCALSINK_USE_THREAD will spawn a thread to process CUDA RPCs instead
of forking a new process. This assumes only one application can be using the
runtime at a time, and that this application must be single-process and
single-threaded.

These will be removed in the future.