Assembly command RPCs enabled.
Developed an RPC protocol for participant ("minion") nodes to communicate with
the central ("main") assembly node. Right now minions can establish a connection
with the main node, and send their static hardware configuration. The main node
builds a list of participants which it uses to match against assembly requests.
The main node spawns an RPC thread which allows remote nodes to connect. They
each receive their own thread on the main node to process assembly RPCs. The
assembly module exports a few functions that enable these threads to update the
participant list needed to maintain a global view of the cluster (listed in
assembly.c under the EXTERNAL FUNCTIONS section). types.h contains data
structures that were originally internal to the assembly module, but which are
now needed to compose assembly RPC messages.

Requests for actual assemblies from applications will also use this RPC medium,
but it has not been implemented yet. Any node will be able to host an
application, and request assemblies. But all requests will be satisfied by the
main node.

I temporarily pulled socket code from Shadowfax to enable development of this
mechanism, while waiting for HATI to become more mature. On top of the socket
code I created an RPC interface layer, in files rpc.c and rpc.h, usable only by
code in assembly.c. This will allow for a different network interface to be used
in the implementation of the RPC mechanism (such as HATI) without affecting any
of the other code.

The executable backend/runtime now takes command arguments, specifying the type
of node it will configure (main or minion) and the location of the main node (if
minion).

Added BUG() in debug.h (just like the one in the kernel). It should/must be used
in areas of the code where state consistency is broken (and is not expected to
be) or conditions that should never exist. A bad example is for return values
from network-based functions, as errors returned by these functions are not
under our control.