Bug fix in shmgrp leader.c
If object B contains struct list_head to link with other object B instances in a
list, and some other object A contains a struct list_head logically used as the
'head' of the list, then A is actually _inside_ the list: N object Bs and 1
object A. When one iterates over the entire list, the iterator pointer will
actually be cast as type object B to storage within object A if *_entry*
functions/macros are used (e.g. list_for_each_entry). Thus, when the loop exits,
one must NOT dereference the iterator unless a boolean is used to indicate the
element was found and that the iterator is valid to dereference.

The alternative is to have object A contain a pointer to object B where struct
list_head entries link object Bs together. The pointer in A to a B object would
indicate whether the list was empty or not.