Remote segv bug fixed on unregFB: use counters
As CUDA applications may call __cudaRegisterFatBinary and the associated
unregister multiple times, we need to keep track of how many there are, so that
we can start and stop processing in the runtime appropriately. The interposer
and remote_sink:minion_threads maintain such counters.

Also lots of other (potential) bug fixes using constructs in list.h