Bug fixes handling symbols passed as strings.
One function in the interposer wasn't fully updated to |= packet flags, instead
resetting it using =. That cleared the CUDA_PKT_SYMB_IS_STRING flag.

The execution layer was dereferencing a structure used to lookup the symbol if
it is an address of an application global instead of a string inside a print
statement.