Multi-threaded apps verified.
Assembly can now actively request some number of GPUs and these will be
satisfied. assemble.c has updated algorithms that, given a request for N GPUs,
it will first try to locate a remote node, then determine if that node has N or
more GPUs. If so, the entire assembly is mapped to that node. Else, it will
default to the local node and again attempt to map N GPUs. If both of these
cases fall through, then an assembly is created with a single GPU mapped locally
for safety (so the code doesn't crash).

Fixed a bug in remotesink.c. If a process has 2+ vGPUs mapped to a remote node,
each acquires a separate thread on that node, associated with each application
thread. When the application terminates, the original thread (which runs main())
will invoke __cudaUnregisterFatBinary. It was observed that the other threads
previously created were not told to stop execution. Thus the main thread now
cancels all threads in remotesink associated with the same application process
on a common host machine.