Fix to cudaHostAlloc.
Use of local- and remotesinks negates the need to RPC this function. The API
documentation says it allocates pinned host memory to enable the GPU to
pull/push with greater bandwidth, knowing the memory is not pageable. Thus
executing it in another process and returning that to the application is
invalid.