Bug fix. Race condition on internals assembly lock in forksink.
Discovered with MPI programs. There's a window between fork and exec where
assembly state is queried. A lock is acquired before accessing this state.
However, it was discovered that this lock somehow was already acquired by
another thread. As forking does not copy any but the invoking thread into the
child, the lock would never be released. Solution was to move the access to this
assembly state and lock (assembly_export) above fork.