Converted runtime to a daemon.
Runtime changes directories to /tmp thus you will need to copy localsink there
for applications to run correctly. This will probably be fixed in some future
modification of the build environment.

All output (stdout and stderr) are redirected to a log file in /tmp as well.
This includes output from all sink processes.