Updated trace code to include exec time.
Each RPC has an additional field specifying the time it spent (as a batch if it
was the last function in it - else it will be zero) executing on the remote
machine. Subtract from the call latency the execution of CUDA functions to
acquire the sum of time spent enqueuing the batch and sending it over the
network.