S3D - remote CUDA Event RPCing enabled
Also enabled cudaStreamDestroy
Fixed other issues with memory functions.

"Refactored" rpc_ops structure to initialize entire struct to DoCudaRPC as it
should be -- adding new functions need not update this structure again, since it
is based on the total size of the struct. Hopefully avoid those bugs in the
future (segfaulting from jumping to address NULL).