added cmdline args and auto launch using PBSENV
main flags
    --master [--pbs]
    --minion=masterDNS
other
    --quiet
    --nohup

Keeneland for some reason sends SIGSTOP to shadowfax when backgrounded:
    shadowfax --master --pbs &
so SIGINT allows it to get some or all minions stopped automatically, but then
cannot continue halting itself, or gets SIGSTOP again. stop_others works but not
when & and SIGSTOP are received. Thus use the scripts added which send SIGINT
and SIGKILL to remote processes.

    $ shadowfax --master --pbs --quiet &
      .. do other things like launch apps ..
    $ tools/scripts/pbs_killall.sh