Bug fix.
marshal.h contained a static array which was included by the interposing
library. Originally there was one C file, but now it is split into categories
(mem, error, stream, ..) each including the file. Each included created a local
copy of the same variable name, making it no longer a global variable as it was
intended.

Solution -- move the storage for the variable into a C file, declare extern in
the header.